+++ src/connector/mysql.rs
@@ -354,8 +354,8 @@ impl Queryable for Mysql {
 #[cfg(test)]
 mod tests {
     use super::MysqlUrl;
+    use crate::tests::test_api::mysql::CONN_STR;
     use crate::{connector::Queryable, error::*, single::Quaint};
-    use crate::{prelude::Select, tests::test_api::mysql::CONN_STR};
     use url::Url;
 
     #[test]
@@ -400,61 +400,5 @@ mod tests {
 
         let err = res.unwrap_err();
         assert!(matches!(err.kind(), ErrorKind::AuthenticationFailed { user } if user == "WRONG"));
-    }
-
-    #[tokio::test]
-    async fn should_pick_up_partially_failed_raw_cmd_scripts() {
-        let conn = Quaint::new(&CONN_STR).await.unwrap();
-
-        let result = conn.raw_cmd("SELECT YOLO; SELECT 1;").await;
-
-        assert!(result.is_err());
-
-        let result = conn.raw_cmd("SELECT 1; SELECT NULL; SELECT YOLO; SELECT 2;").await;
-
-        assert!(result.is_err());
-
-        let error_message = result.unwrap_err().to_string();
-        assert_eq!(error_message, "Error accessing result set, column not found: YOLO");
-    }
-
-    #[tokio::test]
-    async fn should_execute_multi_statement_queries_with_raw_cmd() {
-        let conn = Quaint::new(&CONN_STR).await.unwrap();
-
-        conn.raw_cmd(
-            "
-            CREATE TEMPORARY TABLE `testtable` (id INTEGER PRIMARY KEY);
-            CREATE TEMPORARY TABLE `testtable2` (id INTEGER PRIMARY KEY);
-            INSERT INTO `testtable` (id) VALUES (51);
-            INSERT INTO `testtable2` (id) VALUES (52);
-            ",
-        )
-        .await
-        .unwrap();
-
-        let results = conn
-            .query(Select::from_table("testtable").column("id").into())
-            .await
-            .unwrap();
-
-        let results: Vec<i64> = results
-            .into_iter()
-            .map(|row| row.get("id").unwrap().as_i64().unwrap())
-            .collect();
-
-        assert_eq!(results, &[51]);
-
-        let results = conn
-            .query(Select::from_table("testtable2").column("id").into())
-            .await
-            .unwrap();
-
-        let results: Vec<i64> = results
-            .into_iter()
-            .map(|row| row.get("id").unwrap().as_i64().unwrap())
-            .collect();
-
-        assert_eq!(results, &[52]);
     }
 }
